+++ src/index.js
@@ -3,6 +3,7 @@ import apmonth from './apmonth';
 import apnumber from './apnumber';
 import apstate from './apstate';
 import aptime from './aptime';
+import capfirst from './capfirst';
 import intcomma from './intcomma';
 import intword from './intword';
 import ordinal from './ordinal';
@@ -15,6 +16,7 @@ export {
   apnumber,
   apstate,
   aptime,
+  capfirst,
   intcomma,
   intword,
   ordinal,