+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * Added `aptime` function
+* Added `capfirst` function
 * Added `pluralize` function
 
 ### Removed